move some services from using hibernate daos to spring data jpa repositories

+++ src/main/java/com/openstudies/hibernate/services/common/UserService.java
@@ -20,6 +20,6 @@ public interface UserService {
 
     public User findById(Integer id);
 
-    public void removeById(Long id);
+    public void removeById(Integer id);
 
 }
